auth/requirements.txt
Flask==2.2.3
psycopg2==2.9.5
bcrypt==4.0.1
PyJWT==2.6.0